refactor: 기존의 단일 FirebaseService 를 구독과 공지 전송 2개의 서비스로 분리한다
