+++ src/test/java/com/kustacks/kuring/acceptance/FeedbackAcceptanceTest.java
@@ -1,7 +1,6 @@
 package com.kustacks.kuring.acceptance;
 
 import com.google.firebase.messaging.FirebaseMessagingException;
-import com.kustacks.kuring.message.application.port.in.dto.UserTokenValidationCommand;
 import com.kustacks.kuring.support.IntegrationTestSupport;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -26,7 +25,7 @@ class FeedbackAcceptanceTest extends IntegrationTestSupport {
     @Test
     public void request_feedback() throws FirebaseMessagingException {
         // given
-        doNothing().when(firebaseService).validationToken(anyString());
+        doNothing().when(firebaseSubscribeService).validationToken(anyString());
 
         // when
         var 피드백_요청_응답 = 피드백_요청(USER_FCM_TOKEN, "feedback request");
@@ -39,7 +38,7 @@ class FeedbackAcceptanceTest extends IntegrationTestSupport {
     @Test
     public void request_invalid_length_feedback() throws FirebaseMessagingException {
         // given
-        doNothing().when(firebaseService).validationToken(anyString());
+        doNothing().when(firebaseSubscribeService).validationToken(anyString());
 
         // when
         var 피드백_요청_응답 = 피드백_요청(USER_FCM_TOKEN, "");
