+++ src/test/java/com/kustacks/kuring/worker/update/CategoryNoticeUpdaterTest.java
@@ -1,6 +1,6 @@
 package com.kustacks.kuring.worker.update;
 
-import com.kustacks.kuring.message.application.service.FirebaseService;
+import com.kustacks.kuring.message.application.service.FirebaseNotificationService;
 import com.kustacks.kuring.notice.application.port.out.NoticeQueryPort;
 import com.kustacks.kuring.worker.scrap.KuisNoticeScraperTemplate;
 import com.kustacks.kuring.worker.scrap.client.notice.LibraryNoticeApiClient;
@@ -35,7 +35,7 @@ class CategoryNoticeUpdaterTest {
     KuisNoticeScraperTemplate scrapperTemplate;
 
     @MockBean
-    FirebaseService firebaseService;
+    FirebaseNotificationService firebaseService;
 
     @MockBean
     LibraryNoticeApiClient libraryNoticeApiClient;
